+++ src/types.ts
@@ -1,1 +1,6 @@
 export type SuffixReturn = string | null;
+
+export interface NounObject {
+  single: string;
+  plural: string;
+}
